requirement.txt
python-dotenv
langchain-groq
langchain-core
fastapi
uvicorn
pydantic
python-dateutil